refactor(server): extract the gRPC API into server/apiserver
Move the gRPC management API out of the Server god object into its own package,
split by domain (clients, passwords, connectors, refresh, sessions, identities,
mfa) with the constructor and shared helpers in api.go.

dexAPI held a *Server but used only two things from it: the connector cache (to
invalidate on connector CRUD) and the discovery-document builder. NewAPI now takes
those two narrow dependencies, so apiserver no longer depends on the server
package. The server exposes them through Connectors() and ConstructDiscovery, and
cmd/dex assembles the API — no Server method re-embeds the whole server.

The bcrypt hash-cost check is shared by the API (on password create) and the local
passwordDB connector (on login), so it moves to a small server/passwords package
that both import, rather than being duplicated or forcing a server -> apiserver
dependency.
